refactor(analysis): move core to FSD model segment
domain/ value objects + application/ ports & LRO state machine + the selection
store and InsightSink port now live in src/features/analysis/model. Imports
rewired to intra-segment paths.
