+++ src/features/analysis/model/analysis-error.ts
