+++ src/features/analysis/model/analysis-gateway.ts
@@ -1,5 +1,5 @@
-import type { AnalysisSelection } from "../domain/analysis-selection";
-import type { AnalysisResult } from "../domain/analysis-result";
+import type { AnalysisSelection } from "./analysis-selection";
+import type { AnalysisResult } from "./analysis-result";
 
 /** Opaque reference to a submitted analysis job. */
 export interface JobRef {
