+++ src/features/analysis/model/analysis-result.ts
