+++ src/features/analysis/model/analysis-selection.ts
