+++ src/features/analysis/model/area-selection.ts
