+++ src/features/analysis/model/chart-dto.ts
