+++ src/features/analysis/model/clock.ts
