+++ src/features/analysis/model/insight-sink.ts
